Safari-609.1.11.1/LayoutTests/crypto/subtle/rsassa-pkcs1-v1_5-import-jwk-public-key-sha512-expected.txt

Test importing a JWK RSASSA-PKCS1-v1_5 public key with SHA-512

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'RSASSA-PKCS1-v1_5'
PASS publicKey.algorithm.modulusLength is 2048
PASS bytesToHexString(publicKey.algorithm.publicExponent) is '010001'
PASS publicKey.algorithm.hash.name is 'SHA-512'
PASS publicKey.usages is ['verify']
PASS successfullyParsed is true

TEST COMPLETE
